android: expose the full driver surface through JNI
The JNI shim reached 19 of the driver's 45 public functions and none of
today's additions. The driver itself needed nothing: CMakeLists compiles
../../../../driver/picoscope2204a.c directly, so every change already lands
in the Android build — only the shim was behind.

Adds 12 native methods with their Java facade:

  nativeSetResolutionEnhancement   nativeGetLastOverflow
  nativeSetEts  nativeCaptureEts  nativeDisableEts
  nativeSetTriggerEx  nativeSetTriggerWindow  nativeSetTriggerPwq
  nativeSetSiggenArbitrary         nativeGetStreamingAggregated
  nativeGetEepromCalibration       nativeUseEepromCalibration

Struct returns follow the conventions already in the file: flat primitive
arrays the caller slices. Overflow is int[3]; the EEPROM table is 29 floats
([valid, active, offset×9, gainA×9, gainB×9]); aggregation is 4×buckets
floats laid out minA/maxA/minB/maxB, mirroring nativeGetLatestDual's
A-then-B split; ETS capture uses that same dual layout.

Also fixes a signedness mismatch on ps2204a_set_ets's out-parameter, which
takes int* — it was being handed a uint32_t* in both the JNI shim and the
cgo wrapper.

Verified without an NDK, which turns out to be enough to be sure:

  - the shim compiles under -std=gnu11 with the CMakeLists warning set and
    real JDK jni.h headers, 0 warnings
  - javac -h regenerates the expected JNI symbol names, and cross-checking
    them against the implementations shows 34 declared, 34 implemented,
    nothing orphaned in either direction
  - shim + driver link into a working libpicoscope_jni.so exporting exactly
    those 34 Java_io_github_wasabules_* symbols, with no unresolved
    references beyond libc and libusb

Portability of today's driver work was checked at the same time: no new
libusb calls, only malloc/free/memcpy/pthread, _Static_assert is C11 and
CMAKE_C_STANDARD is already 11, and PS_LOG already routes to
__android_log_print under __ANDROID__.

One latent wart, older than this work: usleep is not declared under strict
-std=c11. CMake defaults to gnu11 so it builds today, but setting
CMAKE_C_EXTENSIONS OFF would break it.

+++ gui/cgo_wrappers.h
@@ -170,10 +170,10 @@ static inline ps_status_t wrap_get_last_overflow(ps2204a_device_t *dev,
 static inline ps_status_t wrap_set_ets(ps2204a_device_t *dev, int mode,
                                        int interleaves, int cycles,
                                        int *out_interval_ps) {
-    uint32_t ips = 0;
+    int ips = 0;
     ps_status_t st = ps2204a_set_ets(dev, (ps_ets_mode_t)mode,
                                      interleaves, cycles, &ips);
-    if (out_interval_ps) *out_interval_ps = (int)ips;
+    if (out_interval_ps) *out_interval_ps = ips;
     return st;
 }
